+++ src/zhdb/主要.py
@@ -1,0 +1,8 @@
+from zhdb.汉字数据库的UserInterface import Ui
+
+def main() -> None:
+    UI = Ui()
+    UI.runMainLoop()
+
+if __name__ == '__main__':
+    main()
